saisie/requirements.txt
numpy
pandas
streamlit
streamlit-pydantic

-e ..
